Seperate sandwich module with sandwich-retrofit
