+++ sandwich-retrofit/src/main/kotlin/com/skydoves/sandwich/retrofit/adapters/internal/ApiResponseDeferredCallAdapter.kt
@@ -13,12 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.skydoves.sandwich.adapters.internal
+package com.skydoves.sandwich.retrofit.adapters.internal
 
 import com.skydoves.sandwich.ApiResponse
+import com.skydoves.sandwich.retrofit.of
+import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Deferred
+import kotlinx.coroutines.launch
 import retrofit2.Call
 import retrofit2.CallAdapter
+import retrofit2.awaitResponse
 import java.lang.reflect.Type
 
 /**
@@ -26,18 +31,38 @@ import java.lang.reflect.Type
  *
  * ApiResponseCallAdapter is an call adapter for creating [ApiResponse] by executing Retrofit's service methods.
  *
- * Request API network call asynchronously and returns [ApiResponse].
+ * Request API network call asynchronously and returns [Deferred] of [ApiResponse].
  */
-internal class ApiResponseCallAdapter constructor(
+internal class ApiResponseDeferredCallAdapter<T>(
   private val resultType: Type,
   private val coroutineScope: CoroutineScope,
-) : CallAdapter<Type, Call<ApiResponse<Type>>> {
+) : CallAdapter<T, Deferred<ApiResponse<T>>> {
 
   override fun responseType(): Type {
     return resultType
   }
 
-  override fun adapt(call: Call<Type>): Call<ApiResponse<Type>> {
-    return ApiResponseCallDelegate(call, coroutineScope)
+  @Suppress("DeferredIsResult")
+  override fun adapt(call: Call<T>): Deferred<ApiResponse<T>> {
+    val deferred = CompletableDeferred<ApiResponse<T>>().apply {
+      invokeOnCompletion {
+        if (isCancelled && !call.isCanceled) {
+          call.cancel()
+        }
+      }
+    }
+
+    coroutineScope.launch {
+      try {
+        val response = call.awaitResponse()
+        val apiResponse = ApiResponse.of { response }
+        deferred.complete(apiResponse)
+      } catch (e: Exception) {
+        val apiResponse = ApiResponse.error<T>(e)
+        deferred.complete(apiResponse)
+      }
+    }
+
+    return deferred
   }
 }
