+++ sandwich-retrofit/src/main/kotlin/com/skydoves/sandwich/retrofit/interceptors/EmptyBodyInterceptor.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.skydoves.sandwich.interceptors
+package com.skydoves.sandwich.retrofit.interceptors
 
 import com.skydoves.sandwich.StatusCode
 import okhttp3.Interceptor
