+++ sandwich/build.gradle.kts
@@ -13,23 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import org.jetbrains.kotlin.gradle.plugin.mpp.KotlinNativeTarget
 
-/*
- * Designed and developed by 2020 skydoves (Jaewoong Eum)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 import com.github.skydoves.sandwich.Configuration
 
 plugins {
@@ -57,10 +41,6 @@ kotlin {
     withJava()
   }
   ios()
-  js(IR) {
-    browser()
-    binaries.executable()
-  }
   iosSimulatorArm64()
   macosArm64()
   macosX64()
@@ -69,33 +49,30 @@ kotlin {
     all { languageSettings.optIn("kotlin.contracts.ExperimentalContracts") }
     val commonMain by getting {
       dependencies {
-        api(libs.retrofit)
+        implementation(libs.coroutines)
+        implementation(libs.okio)
       }
     }
 
     val commonTest by getting {
       dependencies {
+        implementation(libs.coroutines.test)
         implementation(libs.junit)
         implementation(libs.mockito.core)
         implementation(libs.mockito.inline)
         implementation(libs.mockito.kotlin)
-        implementation(libs.retrofit.moshi)
-        implementation(libs.coroutines.test)
         implementation(libs.serialization)
       }
     }
 
     val jvmMain by getting {
       dependencies {
-        api(libs.okhttp)
         implementation(libs.coroutines)
       }
     }
-    val jvmTest by getting {
-      dependencies {
-        implementation(libs.mock.webserver)
-      }
-    }
+
+    val jvmTest by getting
+
     val appleTest by creating {
       dependsOn(commonTest)
     }
