+++ sandwich/src/commonMain/kotlin/com/skydoves/sandwich/SuspensionFunction.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.skydoves.sandwich.adapters.internal
+package com.skydoves.sandwich
 
 /**
  * @author skydoves (Jaewoong Eum)
@@ -21,4 +21,4 @@ package com.skydoves.sandwich.adapters.internal
  * Specifies that this annotation should be used to mark suspension functions.
  */
 @DslMarker
-internal annotation class SuspensionFunction
+public annotation class SuspensionFunction
