+++ settings.gradle.kts
@@ -33,3 +33,4 @@ include(":app")
 include(":sandwich")
 include(":sandwich-datasource")
 include(":sandwich-serialization")
+include(":sandwich-retrofit")
